+++ clodoo/clodoo/examples/import_products_gb.py
@@ -6,7 +6,7 @@ from import_products import (init_n_connect, add_elem)
 # import pdb
 
 
-__version__ = "1.0.1.1"
+__version__ = "1.0.1.2"
 
 MYDICT = {
     'default_code': 0,