+++ clodoo/clodoo/egg-info/history.rst
@@ -1,3 +1,9 @@
+1.0.1.2 (2022-03-29)
+~~~~~~~~~~~~~~~~~~~~
+
+* [FIX] transodoo.xlsx: account.invoice-move
+* [FIX] transodoo.py: account.invoice-move
+
 1.0.1.1 (2022-03-25)
 ~~~~~~~~~~~~~~~~~~~~
 
@@ -5,7 +11,6 @@
 * [IMP] transodoo.py: new features; merge is default type
 * [IMP] list_reqirements.py: minor improvements
 * [IMP] more tests
-
 
 1.0.1 (2022-02-22)
 ~~~~~~~~~~~~~~~~~~~~